requirements.txt
pandas
numpy
scikit-learn
python-dotenv
pymysql
-e .